+++ CHANGELOG.md
@@ -49,6 +49,7 @@ See [0Ver](https://0ver.org/).
 - Adds `Unitable` protocol and `.from_success()` and `.from_failure()`
   methods for all `Result` realted classes
 - Adds `flow` function, which is similar to `pipe`
+- Adds `swap` coverter for `Result` and `IOResult`
 
 
 ### Bugfixes